src/main/java/me/torissi/chapter2/coupling/BankStatementCSVParser2.java
package me.torissi.chapter2.coupling;


import me.torissi.chapter2.srp.BankTransaction;

import java.time.LocalDate;
import java.time.format.DateTimeFormatter;
import java.util.ArrayList;
import java.util.List;

public class BankStatementCSVParser2 implements BankStatementParser {

    final DateTimeFormatter DATE_PATTERN = DateTimeFormatter.ofPattern("yyyy-MM-dd");

    @Override
    public BankTransaction parseFrom(String line) {
        final String[] columns = line.split(",");
        final LocalDate date = LocalDate.parse(columns[0], DATE_PATTERN);
        final double amount = Double.parseDouble(columns[1]);
        final String description = columns[2];

        return new BankTransaction(date, amount, description);
    }

    @Override
    public List<BankTransaction> parseLinesFrom(List<String> lines) {
        final List<BankTransaction> bankTransactions = new ArrayList<>();

        for (final String line: lines) {
            bankTransactions.add(this.parseFrom(line));
        }

        return bankTransactions;
    }
}
